requirements.txt
langchain
langchain-community
langchain-mistralai
python-dotenv
pydantic
